Added in DESim simulation of my simple program.
DESim is a simulation tool that can simulate the behavior of a chip based on a lookup table and a corresponding program, using a memory instantiation file (MIF). With DESim, you can generate an output that shows what the behavior of the chip would be like if it were physically implemented.